New Square Bracket syntax and moving to a snipper
As per forum it is clear that square brackets provide less confusion with the use of liquid. 

The ability to hide the code from external apps has been added. This allows the use of <!--[[shortcode]]-->

This also provides a fix for the RTE surrounding the shortcode in a <p> tag

Also core code has been shifted to a snippet to allow it to be used in multiple places with a single short code